app/requirements.txt
kafka-python==2.0.2
lz4==4.3.2
avro-python3==1.10.2
requests==2.31.0
